requirements.txt
aws-cdk-lib==2.202.0
bedrock-agentcore-starter-toolkit
constructs>=10.0.0,<11.0.0
bedrock-agentcore
strands-agents
strands-agents-tools
